Incomplete encapsolation of sisyphus
