+++ sisyphus/prowstat.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package sisyphus
 
 // ProwResult matches the structure published in finished.json
 type ProwResult struct {
