+++ sisyphus/sisyphus-deployment.yaml
@@ -15,7 +15,7 @@ spec:
         istio: sisyphus
     spec:
       containers:
-      - name: oncall-alert
+      - name: sisyphus-container
         image: gcr.io/istio-testing/sisyphus:0.0.2
         imagePullPolicy: Always
         volumeMounts:
